+++ lunespy/utils/__init__.py
@@ -124,23 +124,3 @@ def lunes_to_unes(lunes: float or int) -> int:
 def unes_to_lunes(unes: int) -> float:
     return float(unes / 10e7)
 
-
-def sha256(object: object) -> str:
-    from hashlib import sha256
-
-    return sha256(
-        str(object).encode()
-    ).hexdigest()
-
-
-def drop_none(data: dict) -> dict:
-    validate_keys = list(filter(
-        lambda key: data[key] != None,
-        data.keys()
-    ))
-
-    return {
-        key: data[key]
-        for key in data.keys()
-        if key in validate_keys
-    }