+++ main.py
@@ -1,0 +1,5 @@
+from lunespy.client.wallet import Account
+
+print(
+    Account(public_key="2uuQVr3B5aGgvSJ5BMCw4Cd19tdYdnMGoYnji99aPde4")
+)